# Conflicts:
#	features/ordermanagement/buyer_order_detail/src/main/java/com/tokopedia/buyerorderdetail/presentation/adapter/BuyerOrderDetailAdapter.kt
#	features/ordermanagement/buyer_order_detail/src/main/java/com/tokopedia/buyerorderdetail/presentation/adapter/typefactory/BuyerOrderDetailTypeFactory.kt
#	features/ordermanagement/buyer_order_detail/src/main/java/com/tokopedia/buyerorderdetail/presentation/fragment/BuyerOrderDetailFragment.kt

+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/domain/response/GetPeopleAddressResponse.kt
@@ -56,7 +56,7 @@ data class ManageAddressData(
         @SerializedName("longitude")
         var longitude: String = "",
         @SerializedName("partner_id")
-        var partnerId: Int = 0,
+        var partnerId: Long = 0,
         @SerializedName("partner_name")
         var partnerName: String = "",
         @SerializedName("phone")
